Updating scripts to check for symlinks.

+++ scripts/createService.sh
@@ -1,4 +1,1 @@
-if [ -f /etc/init.d/id-service ];
-then
-    ln -s /opt/id-service/*.jar /etc/init.d/id-service
-fi
+if [ ! -L /etc/init.d/id-service ]; then ln -s /opt/id-service/*.jar /etc/init.d/id-service; fi
